keyword manager test ver 1.5
명사 추출 갯수 및 영어 명사에 대한 비율도 구한다. 차후에 영어단어가 5할 이상이면 따로 처리할 것.
이를 이용하면 TF를 불린빈도에서 증가빈도로 변경도 용이하게끔 할 수 있을것으로 보인다.

+++ keyword manager/mecabTest2/DB_manager/main.cpp
@@ -7,7 +7,7 @@
 
 using namespace std;
 
-map<pair<string,string>, int> stringTypeNum;
+map<pair<string,string>, double> stringTypeNum;
 
 int main()
 {
@@ -25,7 +25,7 @@ int main()
 	//getline(ImecabTotalTF, ttmp, '\n');
 	//documentN = atoi(ttmp.c_str());
 
-	int documentN;
+	double documentN;
 
 	ImecabTotalTF >> documentN;
 
@@ -87,7 +87,7 @@ int main()
 		ofstream Oresult(mecabTotalTFResult);
 
 		Oresult << endl << documentN << endl;
-		map<pair<string, string>, int>::iterator it;
+		map<pair<string, string>, double>::iterator it;
 		for (it = stringTypeNum.begin(); it != stringTypeNum.end(); ++it)
 			Oresult << it->first.first << '\t' << it->first.second << '\t' << it->second << endl;
